Move implementation of `timestamp*` to `DateTime`
